+++ modules/src/access.rs
@@ -4,6 +4,6 @@ pub mod events;
 mod ownable;
 
 pub use access_control::{
-    AccessControl, AccessControlDeployer, AccessControlRef, Role, DEFAULT_ADMIN_ROLE
+    AccessControl, AccessControlContractRef, AccessControlDeployer, Role, DEFAULT_ADMIN_ROLE
 };
 pub use ownable::*;