+++ modules/src/erc1155/erc1155_base.rs
@@ -1,33 +1,26 @@
 //! Odra module implementing Erc1155 core.
+use super::events::*;
 use crate::erc1155::errors::Error;
-use crate::erc1155::events::{ApprovalForAll, TransferBatch, TransferSingle};
 use crate::erc1155::Erc1155;
-use crate::erc1155_receiver::Erc1155ReceiverRef;
-use odra::types::casper_types::bytesrepr::Bytes;
-use odra::types::OdraAddress;
-use odra::types::U256;
-use odra::{
-    contract_env::{caller, revert},
-    prelude::vec::Vec,
-    types::{event::OdraEvent, Address},
-    Mapping
-};
+use crate::erc1155_receiver::Erc1155ReceiverContractRef;
+use odra::prelude::*;
+use odra::{Address, Bytes, Mapping, Module, OdraAddress, U256};
 
 /// The ERC1155 base implementation.
 #[odra::module(events = [ApprovalForAll, TransferBatch, TransferSingle])]
 pub struct Erc1155Base {
-    pub balances: Mapping<Address, Mapping<U256, U256>>,
-    pub approvals: Mapping<Address, Mapping<Address, bool>>
+    pub balances: Mapping<(Address, U256), U256>,
+    pub approvals: Mapping<(Address, Address), bool>
 }
 
 impl Erc1155 for Erc1155Base {
     fn balance_of(&self, owner: &Address, id: &U256) -> U256 {
-        self.balances.get_instance(owner).get_or_default(id)
+        self.balances.get_or_default(&(*owner, *id))
     }
 
-    fn balance_of_batch(&self, owners: &[Address], ids: &[U256]) -> Vec<U256> {
+    fn balance_of_batch(&self, owners: Vec<Address>, ids: Vec<U256>) -> Vec<U256> {
         if owners.len() != ids.len() {
-            revert(Error::AccountsAndIdsLengthMismatch);
+            self.env().revert(Error::AccountsAndIdsLengthMismatch);
         }
 
         // get balances for each owner and id
@@ -41,23 +34,22 @@ impl Erc1155 for Erc1155Base {
     }
 
     fn set_approval_for_all(&mut self, operator: &Address, approved: bool) {
-        let owner = self.caller();
-        if owner == operator {
-            revert(Error::ApprovalForSelf);
+        let owner = self.env().caller();
+        if owner == *operator {
+            self.env().revert(Error::ApprovalForSelf);
         }
 
-        self.approvals.get_instance(owner).set(operator, approved);
+        self.approvals.set(&(owner, *operator), approved);
 
-        ApprovalForAll {
-            owner: *owner,
+        self.env().emit_event(ApprovalForAll {
+            owner,
             operator: *operator,
             approved
-        }
-        .emit();
+        });
     }
 
     fn is_approved_for_all(&self, owner: &Address, operator: &Address) -> bool {
-        self.approvals.get_instance(owner).get_or_default(operator)
+        self.approvals.get_or_default(&(*owner, *operator))
     }
 
     fn safe_transfer_from(
@@ -68,29 +60,25 @@ impl Erc1155 for Erc1155Base {
         amount: &U256,
         data: &Option<Bytes>
     ) {
-        let caller = caller();
+        let caller = self.env().caller();
         self.assert_approved_or_owner(&caller, from);
 
         let from_balance = self.balance_of(from, id);
         if from_balance < *amount {
-            revert(Error::InsufficientBalance);
+            self.env().revert(Error::InsufficientBalance);
         }
 
+        self.balances.set(&(*from, *id), from_balance - *amount);
         self.balances
-            .get_instance(from)
-            .set(id, from_balance - *amount);
-        self.balances
-            .get_instance(to)
-            .set(id, self.balance_of(to, id) + *amount);
+            .set(&(*to, *id), self.balance_of(to, id) + *amount);
 
-        TransferSingle {
+        self.env().emit_event(TransferSingle {
             operator: Some(caller),
             from: Some(*from),
             to: Some(*to),
             id: *id,
             value: *amount
-        }
-        .emit();
+        });
 
         // verify the recipient
         self.safe_transfer_acceptance_check(&caller, from, to, id, amount, data);
@@ -100,15 +88,15 @@ impl Erc1155 for Erc1155Base {
         &mut self,
         from: &Address,
         to: &Address,
-        ids: &[U256],
-        amounts: &[U256],
+        ids: Vec<U256>,
+        amounts: Vec<U256>,
         data: &Option<Bytes>
     ) {
-        let caller = caller();
+        let caller = self.env().caller();
         self.assert_approved_or_owner(&caller, from);
 
         if ids.len() != amounts.len() {
-            revert(Error::IdsAndAmountsLengthMismatch);
+            self.env().revert(Error::IdsAndAmountsLengthMismatch);
         }
 
         // batch transfer
@@ -119,26 +107,22 @@ impl Erc1155 for Erc1155Base {
             // balance check - if a single transfer is incorrect the whole transaction reverts.
             let from_balance = self.balance_of(from, &id);
             if from_balance < amount {
-                revert(Error::InsufficientBalance);
+                self.env().revert(Error::InsufficientBalance);
             }
 
             // update balance
+            self.balances.set(&(*from, id), from_balance - amount);
             self.balances
-                .get_instance(from)
-                .set(&id, from_balance - amount);
-            self.balances
-                .get_instance(to)
-                .set(&id, self.balance_of(to, &id) + amount);
+                .set(&(*to, id), self.balance_of(to, &id) + amount);
         }
 
-        TransferBatch {
+        self.env().emit_event(TransferBatch {
             operator: Some(caller),
             from: Some(*from),
             to: Some(*to),
             ids: ids.to_vec(),
             values: amounts.to_vec()
-        }
-        .emit();
+        });
 
         // verify the recipient
         self.safe_batch_transfer_acceptance_check(&caller, from, to, ids, amounts, data);
@@ -155,7 +139,7 @@ impl Erc1155Base {
 
     fn assert_approved_or_owner(&self, spender: &Address, owner: &Address) {
         if !self.is_approved_or_owner(spender, owner) {
-            revert(Error::NotAnOwnerOrApproved);
+            self.env().revert(Error::NotAnOwnerOrApproved);
         }
     }
 
@@ -171,10 +155,13 @@ impl Erc1155Base {
         data: &Option<Bytes>
     ) {
         if to.is_contract() {
-            let response =
-                Erc1155ReceiverRef::at(to).on_erc1155_received(operator, from, id, amount, data);
+            let response = Erc1155ReceiverContractRef {
+                env: self.env(),
+                address: *to
+            }
+            .on_erc1155_received(*operator, *from, *id, *amount, data.clone());
             if !response {
-                revert(Error::TransferRejected);
+                self.env().revert(Error::TransferRejected);
             }
         }
     }
@@ -186,15 +173,24 @@ impl Erc1155Base {
         operator: &Address,
         from: &Address,
         to: &Address,
-        ids: &[U256],
-        amounts: &[U256],
+        ids: Vec<U256>,
+        amounts: Vec<U256>,
         data: &Option<Bytes>
     ) {
         if to.is_contract() {
-            let response = Erc1155ReceiverRef::at(to)
-                .on_erc1155_batch_received(operator, from, ids, amounts, data);
+            let response = Erc1155ReceiverContractRef {
+                env: self.env(),
+                address: *to
+            }
+            .on_erc1155_batch_received(
+                *operator,
+                *from,
+                ids.to_vec(),
+                amounts.to_vec(),
+                data.clone()
+            );
             if !response {
-                revert(Error::TransferRejected);
+                self.env().revert(Error::TransferRejected);
             }
         }
     }