+++ modules/src/erc721/extensions/erc721_metadata.rs
@@ -1,8 +1,11 @@
 //! Erc721 metadata.
 
+use crate::erc721::extensions::erc721_metadata::errors::Error::{
+    BaseUriNotSet, NameNotSet, SymbolNotSet
+};
 use odra::prelude::*;
+use odra::UnwrapOrRevert;
 use odra::{Module, Variable};
-use crate::erc721::extensions::erc721_metadata::errors::Error::{BaseUriNotSet, NameNotSet, SymbolNotSet};
 
 /// The ERC721 Metadata interface as defined in the standard.
 pub trait Erc721Metadata {
@@ -23,17 +26,21 @@ pub struct Erc721MetadataExtension {
 
 impl Erc721Metadata for Erc721MetadataExtension {
     fn name(&self) -> String {
-        self.name.get().unwrap_or_revert_with(self.env(), NameNotSet)
+        self.name
+            .get()
+            .unwrap_or_revert_with(&self.env(), NameNotSet)
     }
 
     fn symbol(&self) -> String {
-        self.symbol.get().unwrap_or_revert_with(self.env(), SymbolNotSet)
+        self.symbol
+            .get()
+            .unwrap_or_revert_with(&self.env(), SymbolNotSet)
     }
 
     fn base_uri(&self) -> String {
         self.base_uri
             .get()
-            .unwrap_or_revert_with(self.env(), BaseUriNotSet)
+            .unwrap_or_revert_with(&self.env(), BaseUriNotSet)
     }
 }
 
@@ -49,12 +56,10 @@ pub mod errors {
     use odra::OdraError;
 
     pub enum Error {
-            NameNotSet = 31_000,
-            SymbolNotSet = 31_001,
-            BaseUriNotSet = 31_002,
-        }
-
-
+        NameNotSet = 31_000,
+        SymbolNotSet = 31_001,
+        BaseUriNotSet = 31_002
+    }
 
     impl From<Error> for OdraError {
         fn from(error: Error) -> Self {